requirements.txt
mesop
ollama
groq
python-dotenv
